requirements.txt
flask>=2.0.0
Pillow>=8.0.0
numpy>=1.19.0
requests>=2.25.0
scipy>=1.5.0
python-dotenv>=0.19.0 